# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  imageUrl?: string | null;
}

export type ProjectRole = (typeof ProjectRole)[keyof typeof ProjectRole];

export const ProjectRole = {
  admin: "admin",
  member: "member",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Project {
  id: string;
  name: string;
  description?: string | null;
  color: string;
  ownerId: string;
  createdAt: string;
}

export interface ProjectSummary {
  id: string;
  name: string;
  description?: string | null;
  color: string;
  ownerId: string;
  createdAt: string;
  myRole: ProjectRole;
  memberCount: number;
  totalTasks: number;
  todoCount: number;
  inProgressCount: number;
  doneCount: number;
  overdueCount: number;
}

export interface ProjectMember {
  id: string;
  projectId: string;
  userId: string;
  role: ProjectRole;
  joinedAt: string;
  user: User;
}

export type ProjectDetail = ProjectSummary & {
  members: ProjectMember[];
};

export interface CreateProjectInput {
  /**
   * @minLength 1
   * @maxLength 80
   */
  name: string;
  /** @maxLength 500 */
  description?: string | null;
  color: string;
}

export interface UpdateProjectInput {
  /**
   * @minLength 1
   * @maxLength 80
   */
  name?: string;
  /** @maxLength 500 */
  description?: string | null;
  color?: string;
}

export interface AddMemberInput {
  userId: string;
  role: ProjectRole;
}

export interface UpdateMemberInput {
  role: ProjectRole;
}

export interface Task {
  id: string;
  projectId: string;
  title: string;
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  assigneeId?: string | null;
  assignee?: User | null;
  dueDate?: string | null;
  createdById: string;
  createdAt: string;
  updatedAt: string;
}

export type TaskWithProject = Task & {
  project: Project;
};

export interface CreateTaskInput {
  /**
   * @minLength 1
   * @maxLength 140
   */
  title: string;
  /** @maxLength 2000 */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  assigneeId?: string | null;
  dueDate?: string | null;
}

export interface UpdateTaskInput {
  /**
   * @minLength 1
   * @maxLength 140
   */
  title?: string;
  /** @maxLength 2000 */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  assigneeId?: string | null;
  dueDate?: string | null;
}

export interface StatusBucket {
  status: TaskStatus;
  count: number;
}

export interface Dashboard {
  totalProjects: number;
  totalTasks: number;
  myOpenTasks: number;
  myOverdueTasks: number;
  myCompletedThisWeek: number;
  statusBreakdown: StatusBucket[];
  upcomingTasks: TaskWithProject[];
  overdueTasks: TaskWithProject[];
}

export type ActivityEntryKind =
  (typeof ActivityEntryKind)[keyof typeof ActivityEntryKind];

export const ActivityEntryKind = {
  task_created: "task_created",
  task_status_changed: "task_status_changed",
  task_assigned: "task_assigned",
  project_created: "project_created",
  member_added: "member_added",
} as const;

export interface ActivityEntry {
  id: string;
  kind: ActivityEntryKind;
  projectId: string;
  projectName: string;
  actorId: string;
  actorName: string;
  targetTaskId?: string | null;
  targetTaskTitle?: string | null;
  message: string;
  createdAt: string;
}