# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects the current user can access
      responses:
        "200":
          description: Projects with summary stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (creator becomes Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (Admin only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (Admin only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addProjectMember
      tags: [members]
      summary: Add a user to a project (Admin only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/members/{memberId}:
    patch:
      operationId: updateProjectMember
      tags: [members]
      summary: Change a member's role (Admin only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
    delete:
      operationId: removeProjectMember
      tags: [members]
      summary: Remove a member from a project (Admin only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        imageUrl: { type: string, nullable: true }
      required: [id, email, name]
    ProjectRole:
      type: string
      enum: [admin, member]
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        color: { type: string }
        ownerId: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, name, color, ownerId, createdAt]
    ProjectSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        color: { type: string }
        ownerId: { type: string }
        createdAt: { type: string, format: date-time }
        myRole:
          $ref: "#/components/schemas/ProjectRole"
        memberCount: { type: integer }
        totalTasks: { type: integer }
        todoCount: { type: integer }
        inProgressCount: { type: integer }
        doneCount: { type: integer }
        overdueCount: { type: integer }
      required:
        - id
        - name
        - color
        - ownerId
        - createdAt
        - myRole
        - memberCount
        - totalTasks
        - todoCount
        - inProgressCount
        - doneCount
        - overdueCount
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/ProjectMember"
          required: [members]
    CreateProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500, nullable: true }
        color: { type: string }
      required: [name, color]
    UpdateProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500, nullable: true }
        color: { type: string }
    ProjectMember:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        role:
          $ref: "#/components/schemas/ProjectRole"
        joinedAt: { type: string, format: date-time }
        user:
          $ref: "#/components/schemas/User"
      required: [id, projectId, userId, role, joinedAt, user]
    AddMemberInput:
      type: object
      properties:
        userId: { type: string }
        role:
          $ref: "#/components/schemas/ProjectRole"
      required: [userId, role]
    UpdateMemberInput:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/ProjectRole"
      required: [role]
...