# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  /**
   * @minLength 1
   * @maxLength 80
   */
  name: string;
  /** @maxLength 500 */
  description?: string | null;
  color: string;
}