# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectMember } from "./projectMember";
import type { ProjectSummary } from "./projectSummary";

export type ProjectDetail = ProjectSummary & {
  members: ProjectMember[];
};